reorganize things a bit
