+++ fns-csv-pipeline/src/main/kotlin/gov/cdc/dex/csv/services/AzureBlobServiceImpl.kt
@@ -7,11 +7,7 @@ import com.azure.storage.blob.BlobServiceClientBuilder
 import java.io.InputStream
 import java.io.OutputStream
 
-class AzureBlobServiceImpl(connectionStr: String) :BlobService {
-    private val ingestBlobContainer = System.getenv("BlobContainer_Ingest")
-    private val processedBlobContainer = System.getenv("BlobContainer_Processed")
-    private val errorBlobContainer = System.getenv("BlobContainer_Error")
-
+class AzureBlobServiceImpl(connectionStr:String, ingestBlobContainer:String, processedBlobContainer:String, errorBlobContainer:String) :BlobService {
     private val blobServiceClient: BlobServiceClient = BlobServiceClientBuilder().connectionString(connectionStr).buildClient();
     private val ingestClient: BlobContainerClient = blobServiceClient.getBlobContainerClient(ingestBlobContainer);
     private val processedClient: BlobContainerClient = blobServiceClient.getBlobContainerClient(processedBlobContainer);
